+++ website/src/constants.js
@@ -12,7 +12,7 @@ export const API = {
     SIGN_IN_WITH_EMAIL: 'auth/sign-in-with-email',
     SIGN_IN_WITH_SOCIAL_MEDIA: 'auth/sign-in-with-social-media',
     AUTH_WITH_SOCIAL_MEDIA: 'auth/auth-with-social-media',
-    UNIVERSAL: 'universal',
+    UNIVERSAL_SOCIAL_MEDIA_TYPE: 'universal/social-media-types',
     NUMBER_OF_REGISTERED_USERS: 'users/number-of-registered-users',
     CHECK_IN: 'users/check-in'
 };
