+++ website/src/template/sign-in.html
@@ -60,17 +60,17 @@
                         </div>
 
                         <div class="d-flex flex-column font-weight-bold">
-                            <button tabindex="4"
+                            <button onclick="controller.signInWithFacebook" tabindex="4"
                                 class="btn btn-md btn-primary rounded-pill h4 mb-3 d-flex align-items-center justify-content-center py-25">
                                 <img src="/assets/images/facebook.svg" alt="facebook icon" class="me-2" />
                                 使用 FACEBOOK 登入
                             </button>
-                            <button tabindex="5"
+                            <button onclick="controller.signInWithGoogle" tabindex="5"
                                 class="btn btn-md btn-primary rounded-pill h4 mb-3 d-flex align-items-center justify-content-center py-25">
                                 <img src="/assets/images/google.svg" alt="google icon" class="me-2" />
                                 使用 GOOGLE 登入
                             </button>
-                            <button tabindex="6"
+                            <button onclick="controller.signInWithLine" tabindex="6"
                                 class="btn btn-md btn-primary rounded-pill h4 mb-3 d-flex align-items-center justify-content-center py-25">
                                 <img src="/assets/images/line.svg" alt="line icon" class="me-2" />
                                 使用 LINE 登入
